refactor: renamed and rearrange code in git package
