+++ internal/git/file.go
@@ -17,16 +17,10 @@ package git
 
 import (
 	"bufio"
-	"fmt"
 	"os"
 	"path/filepath"
 
 	"github.com/go-git/go-billy/v5"
-	"github.com/go-git/go-billy/v5/memfs"
-	"github.com/go-git/go-git/v5"
-	"github.com/go-git/go-git/v5/storage"
-	"github.com/go-git/go-git/v5/storage/memory"
-	"github.com/mia-platform/vab/pkg/apis/vab.mia-platform.eu/v1alpha1"
 )
 
 // File rappresent a file downloaded in the in memory store
@@ -62,38 +56,4 @@ func (f *File) WriteContent(targetPath string) error {
 
 	_, err = r.WriteTo(w)
 	return err
-}
-
-type FilesGetter struct {
-	fs           billy.Filesystem
-	storage      *memory.Storage
-	clonePackage func(billy.Filesystem, storage.Storer, v1alpha1.Package) (billy.Filesystem, error)
-}
-
-// NewFilesGetter create a new FilesGetter instance configured for downloading from remote repository using
-// an in memory storage
-func NewFilesGetter() *FilesGetter {
-	return &FilesGetter{
-		fs:      memfs.New(),
-		storage: memory.NewStorage(),
-		clonePackage: func(fs billy.Filesystem, storage storage.Storer, pkg v1alpha1.Package) (billy.Filesystem, error) {
-			cloneOptions := cloneOptionsForPackage(pkg)
-			if _, err := git.Clone(storage, fs, cloneOptions); err != nil {
-				return nil, fmt.Errorf("error cloning repository %w", err)
-			}
-
-			return fs, nil
-		},
-	}
-}
-
-// GetFilesForPackage clones the pkg from the remote repository and return all the files relative for the package
-// or an error otherwise
-func (r *FilesGetter) GetFilesForPackage(pkg v1alpha1.Package) ([]*File, error) {
-	memFs, err := r.clonePackage(r.fs, r.storage, pkg)
-	if err != nil {
-		return nil, err
-	}
-
-	return filterWorktreeForPackage(memFs, pkg)
 }
